add observability calls, move AsyncCallQueueDepth to async_queries.sql
